localRating is working!

+++ The_Reviewer_Project/src/main/java/academy/mindswap/persistence/models/Review.java
@@ -37,7 +37,7 @@ public class Review {
     private String description;
 
     @Column
-    private Integer localRating;
+    private float localRating;
 
     @Override
     public boolean equals(Object o) {
